+++ ThirdPartyAdapters/dtexchange/dtexchange/src/test/kotlin/com/google/ads/mediation/fyber/FyberMediationAdapterTest.kt
@@ -19,13 +19,19 @@ import androidx.core.os.bundleOf
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.fyber.inneractive.sdk.external.BidTokenProvider
 import com.fyber.inneractive.sdk.external.InneractiveAdManager
+import com.fyber.inneractive.sdk.external.InneractiveAdRequest
 import com.fyber.inneractive.sdk.external.InneractiveAdSpot
 import com.fyber.inneractive.sdk.external.InneractiveAdSpotManager
+import com.fyber.inneractive.sdk.external.InneractiveFullscreenUnitController
+import com.fyber.inneractive.sdk.external.InneractiveMediationDefs
+import com.fyber.inneractive.sdk.external.InneractiveUserConfig
 import com.fyber.inneractive.sdk.external.OnFyberMarketplaceInitializedListener
 import com.fyber.inneractive.sdk.external.OnFyberMarketplaceInitializedListener.FyberInitStatus
 import com.google.ads.mediation.adaptertestkit.AdErrorMatcher
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants
+import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_APP_ID
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_BID_RESPONSE
+import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_PLACEMENT_ID
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_WATERMARK
 import com.google.ads.mediation.adaptertestkit.assertGetSdkVersion
 import com.google.ads.mediation.adaptertestkit.assertGetVersionInfo
@@ -34,8 +40,10 @@ import com.google.ads.mediation.adaptertestkit.createMediationConfiguration
 import com.google.ads.mediation.adaptertestkit.createMediationInterstitialAdConfiguration
 import com.google.ads.mediation.adaptertestkit.createMediationNativeAdConfiguration
 import com.google.ads.mediation.adaptertestkit.createMediationRewardedAdConfiguration
+import com.google.ads.mediation.fyber.DTExchangeErrorCodes.ERROR_INVALID_SERVER_PARAMETERS
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.AdFormat
+import com.google.android.gms.ads.AgeRestrictedTreatment
 import com.google.android.gms.ads.MobileAds
 import com.google.android.gms.ads.RequestConfiguration
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback
@@ -50,6 +58,7 @@ import com.google.android.gms.ads.mediation.MediationRewardedAdCallback
 import com.google.android.gms.ads.mediation.NativeAdMapper
 import com.google.android.gms.ads.mediation.rtb.RtbSignalData
 import com.google.android.gms.ads.mediation.rtb.SignalCallbacks
+import com.google.common.truth.Truth.assertThat
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -77,8 +86,8 @@ class FyberMediationAdapterTest {
 
   private val serverParameters =
     bundleOf(
-      FyberMediationAdapter.KEY_APP_ID to AdapterTestKitConstants.TEST_APP_ID,
-      FyberMediationAdapter.KEY_SPOT_ID to AdapterTestKitConstants.TEST_PLACEMENT_ID,
+      FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID,
+      FyberMediationAdapter.KEY_SPOT_ID to TEST_PLACEMENT_ID,
     )
   private val mockInitializationCompleteCallback: InitializationCompleteCallback = mock()
   private val mockBannerAdLoadCallback:
@@ -106,6 +115,7 @@ class FyberMediationAdapterTest {
           RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
         )
         .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .setAgeRestrictedTreatment(AgeRestrictedTreatment.UNSPECIFIED)
         .build()
     MobileAds.setRequestConfiguration(requestConfiguration)
     adapter = FyberMediationAdapter()
@@ -174,6 +184,30 @@ class FyberMediationAdapterTest {
     val requestConfiguration =
       RequestConfiguration.Builder()
         .setTagForChildDirectedTreatment(RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
+    val initializationParameters = createMediationConfiguration(AdFormat.BANNER, serverParameters)
+    adapter.initialize(
+      activity,
+      mockInitializationCompleteCallback,
+      listOf(initializationParameters),
+    )
+
+    mockInneractiveAdManager.verify {
+      InneractiveAdManager.currentAudienceAppliesToCoppa()
+      InneractiveAdManager.initialize(eq(activity), eq(AdapterTestKitConstants.TEST_APP_ID), any())
+    }
+  }
+
+  @Test
+  fun initialize_withAgeRestrictedTreatmentChild_initializesInneractiveAdManagerAndSetsCoppa() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .setAgeRestrictedTreatment(AgeRestrictedTreatment.CHILD)
         .build()
     MobileAds.setRequestConfiguration(requestConfiguration)
     val initializationParameters = createMediationConfiguration(AdFormat.BANNER, serverParameters)
@@ -332,7 +366,139 @@ class FyberMediationAdapterTest {
 
   // endregion
 
-  // region Banner Ad Load Tests
+  // region Waterfall Banner Ad load tests
+
+  @Test
+  fun loadBannerAd_ifAppIdIsMissing_fails() {
+    adapter.loadBannerAd(
+      createMediationBannerAdConfiguration(context = activity),
+      mockBannerAdLoadCallback,
+    )
+
+    val adErrorCaptor = argumentCaptor<AdError>()
+    verify(mockBannerAdLoadCallback).onFailure(adErrorCaptor.capture())
+    val adError = adErrorCaptor.firstValue
+    assertThat(adError.code).isEqualTo(ERROR_INVALID_SERVER_PARAMETERS)
+    assertThat(adError.domain).isEqualTo(DTExchangeErrorCodes.ERROR_DOMAIN)
+  }
+
+  @Test
+  fun loadBannerAd_ifSpotIdIsMissing_fails() {
+    adapter.loadBannerAd(
+      createMediationBannerAdConfiguration(
+        context = activity,
+        serverParameters = bundleOf(FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID_1),
+      ),
+      mockBannerAdLoadCallback,
+    )
+
+    val adErrorCaptor = argumentCaptor<AdError>()
+    verify(mockBannerAdLoadCallback).onFailure(adErrorCaptor.capture())
+    val adError = adErrorCaptor.firstValue
+    assertThat(adError.code).isEqualTo(ERROR_INVALID_SERVER_PARAMETERS)
+    assertThat(adError.domain).isEqualTo(DTExchangeErrorCodes.ERROR_DOMAIN)
+  }
+
+  @Test
+  fun loadBannerAd_setsMediationNameAndVersionAndInitializesDtSdk() {
+    adapter.loadBannerAd(
+      createMediationBannerAdConfiguration(
+        context = activity,
+        serverParameters =
+          bundleOf(
+            FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID_1,
+            FyberMediationAdapter.KEY_SPOT_ID to TEST_SPOT_ID,
+          ),
+      ),
+      mockBannerAdLoadCallback,
+    )
+
+    mockInneractiveAdManager.verify {
+      InneractiveAdManager.setMediationName(FyberMediationAdapter.MEDIATOR_NAME)
+    }
+    mockInneractiveAdManager.verify { InneractiveAdManager.setMediationVersion(any()) }
+    mockInneractiveAdManager.verify {
+      InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID_1), any())
+    }
+  }
+
+  @Test
+  fun loadBannerAd_ifInitFails_fails() {
+    adapter.loadBannerAd(
+      createMediationBannerAdConfiguration(
+        context = activity,
+        serverParameters =
+          bundleOf(
+            FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID_1,
+            FyberMediationAdapter.KEY_SPOT_ID to TEST_SPOT_ID,
+          ),
+      ),
+      mockBannerAdLoadCallback,
+    )
+    val initListenerCaptor = argumentCaptor<OnFyberMarketplaceInitializedListener>()
+    mockInneractiveAdManager.verify {
+      InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID_1), initListenerCaptor.capture())
+    }
+    initListenerCaptor.firstValue.onFyberMarketplaceInitialized(FyberInitStatus.FAILED)
+
+    val expectedAdError =
+      AdError(
+        202,
+        "DT Exchange failed to initialize with reason: ${FyberInitStatus.FAILED}",
+        DTExchangeErrorCodes.ERROR_DOMAIN,
+      )
+    verify(mockBannerAdLoadCallback).onFailure(argThat(AdErrorMatcher(expectedAdError)))
+  }
+
+  @Test
+  fun loadBannerAd_ifInitSucceeds_loadsBannerAd() {
+    mockStatic(InneractiveAdSpotManager::class.java).use {
+      val bannerSpot = mock<InneractiveAdSpot>()
+      val mockInneractiveAdSpotManager =
+        mock<InneractiveAdSpotManager> { on { createSpot() } doReturn bannerSpot }
+      whenever(InneractiveAdSpotManager.get()) doReturn mockInneractiveAdSpotManager
+
+      adapter.loadBannerAd(
+        createMediationBannerAdConfiguration(
+          context = activity,
+          serverParameters =
+            bundleOf(
+              FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID_1,
+              FyberMediationAdapter.KEY_SPOT_ID to TEST_SPOT_ID,
+            ),
+          mediationExtras =
+            bundleOf(
+              InneractiveMediationDefs.KEY_AGE to 20,
+              FyberMediationAdapter.KEY_MUTE_VIDEO to true,
+            ),
+        ),
+        mockBannerAdLoadCallback,
+      )
+      val initListenerCaptor = argumentCaptor<OnFyberMarketplaceInitializedListener>()
+      mockInneractiveAdManager.verify {
+        InneractiveAdManager.initialize(
+          eq(activity),
+          eq(TEST_APP_ID_1),
+          initListenerCaptor.capture(),
+        )
+      }
+      initListenerCaptor.firstValue.onFyberMarketplaceInitialized(FyberInitStatus.SUCCESSFULLY)
+
+      verify(bannerSpot).addUnitController(any())
+      verify(bannerSpot).setRequestListener(any())
+      val inneractiveAdRequestCaptor = argumentCaptor<InneractiveAdRequest>()
+      verify(bannerSpot).requestAd(inneractiveAdRequestCaptor.capture())
+      assertThat(inneractiveAdRequestCaptor.firstValue.spotId).isEqualTo(TEST_SPOT_ID)
+      val userParamsCaptor = argumentCaptor<InneractiveUserConfig>()
+      mockInneractiveAdManager.verify {
+        InneractiveAdManager.setUserParams(userParamsCaptor.capture())
+      }
+      assertThat(userParamsCaptor.firstValue.age).isEqualTo(20)
+      mockInneractiveAdManager.verify { InneractiveAdManager.setMuteVideo(true) }
+    }
+  }
+
+  // region RTB Banner Ad Load Tests
   @Test
   fun loadRtbBannerAd_invokesLoadAd() {
     mockStatic(InneractiveAdSpotManager::class.java).use {
@@ -359,7 +525,122 @@ class FyberMediationAdapterTest {
 
   // endregion
 
-  // region Banner Ad Load Tests
+  // region Waterfall Interstitial Ad Load tests
+
+  @Test
+  fun loadInterstitialAd_withEmptyAppId_fails() {
+    val invalidServerParameters = bundleOf(FyberMediationAdapter.KEY_SPOT_ID to TEST_SPOT_ID)
+    val adConfiguration =
+      createMediationInterstitialAdConfiguration(
+        context = activity,
+        serverParameters = invalidServerParameters,
+      )
+
+    adapter.loadInterstitialAd(adConfiguration, mockInterstitialAdLoadCallback)
+
+    val expectedAdError =
+      AdError(
+        ERROR_INVALID_SERVER_PARAMETERS,
+        "App ID is null or empty.",
+        DTExchangeErrorCodes.ERROR_DOMAIN,
+      )
+    verify(mockInterstitialAdLoadCallback).onFailure(argThat(AdErrorMatcher(expectedAdError)))
+  }
+
+  @Test
+  fun loadInterstitialAd_withEmptySpotId_fails() {
+    val invalidServerParameters = bundleOf(FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID)
+    val adConfiguration =
+      createMediationInterstitialAdConfiguration(
+        context = activity,
+        serverParameters = invalidServerParameters,
+      )
+
+    adapter.loadInterstitialAd(adConfiguration, mockInterstitialAdLoadCallback)
+
+    val expectedAdError =
+      AdError(
+        ERROR_INVALID_SERVER_PARAMETERS,
+        "Cannot render interstitial ad. Please define a valid spot id on the AdMob UI.",
+        DTExchangeErrorCodes.ERROR_DOMAIN,
+      )
+    verify(mockInterstitialAdLoadCallback).onFailure(argThat(AdErrorMatcher(expectedAdError)))
+  }
+
+  @Test
+  fun loadInterstitialAd_initializesInneractiveAdManager() {
+    val adConfiguration =
+      createMediationInterstitialAdConfiguration(
+        context = activity,
+        serverParameters = serverParameters,
+      )
+
+    adapter.loadInterstitialAd(adConfiguration, mockInterstitialAdLoadCallback)
+
+    mockInneractiveAdManager.verify {
+      InneractiveAdManager.setMediationName(FyberMediationAdapter.MEDIATOR_NAME)
+    }
+    mockInneractiveAdManager.verify { InneractiveAdManager.setMediationVersion(any()) }
+    mockInneractiveAdManager.verify {
+      InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID), any())
+    }
+  }
+
+  @Test
+  fun loadInterstitialAd_whenInitFails_fails() {
+    val adConfiguration =
+      createMediationInterstitialAdConfiguration(
+        context = activity,
+        serverParameters = serverParameters,
+      )
+    val listenerCaptor = argumentCaptor<OnFyberMarketplaceInitializedListener>()
+
+    adapter.loadInterstitialAd(adConfiguration, mockInterstitialAdLoadCallback)
+    mockInneractiveAdManager.verify {
+      InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID), listenerCaptor.capture())
+    }
+    listenerCaptor.firstValue.onFyberMarketplaceInitialized(FyberInitStatus.FAILED)
+
+    val expectedAdError =
+      AdError(
+        202,
+        "DT Exchange failed to initialize with reason: FAILED",
+        DTExchangeErrorCodes.ERROR_DOMAIN,
+      )
+    verify(mockInterstitialAdLoadCallback).onFailure(argThat(AdErrorMatcher(expectedAdError)))
+  }
+
+  @Test
+  fun loadInterstitialAd_whenInitSucceeds_requestsAd() {
+    mockStatic(InneractiveAdSpotManager::class.java).use {
+      val interstitialSpot = mock<InneractiveAdSpot>()
+      val mockInneractiveAdSpotManager =
+        mock<InneractiveAdSpotManager> { on { createSpot() } doReturn interstitialSpot }
+      whenever(InneractiveAdSpotManager.get()) doReturn mockInneractiveAdSpotManager
+      val adConfiguration =
+        createMediationInterstitialAdConfiguration(
+          context = activity,
+          serverParameters = serverParameters,
+        )
+      val listenerCaptor = argumentCaptor<OnFyberMarketplaceInitializedListener>()
+
+      adapter.loadInterstitialAd(adConfiguration, mockInterstitialAdLoadCallback)
+      mockInneractiveAdManager.verify {
+        InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID), listenerCaptor.capture())
+      }
+      listenerCaptor.firstValue.onFyberMarketplaceInitialized(FyberInitStatus.SUCCESSFULLY)
+
+      verify(interstitialSpot).addUnitController(any<InneractiveFullscreenUnitController>())
+      verify(interstitialSpot).setRequestListener(any())
+      val requestCaptor = argumentCaptor<InneractiveAdRequest>()
+      verify(interstitialSpot).requestAd(requestCaptor.capture())
+      assertThat(requestCaptor.firstValue.spotId).isEqualTo(TEST_PLACEMENT_ID)
+    }
+  }
+
+  // endregion
+
+  // region RTB Interstitial Ad Load Tests
   @Test
   fun loadRtbInterstitialAd_invokesLoadAd() {
     mockStatic(InneractiveAdSpotManager::class.java).use {
@@ -522,5 +803,6 @@ class FyberMediationAdapterTest {
   private companion object {
     const val TEST_APP_ID_1 = "testAppID1"
     const val TEST_APP_ID_2 = "testAppID2"
+    const val TEST_SPOT_ID = "testSpotId"
   }
 }